Death Vignette is in!
Circle shrinks based on remaining lifeForce of the player and is centered on them as well. The circle also grows if the player absorbs lifeForce from plants and shrinks if the player imbues life to plants.

Currently the center of the circle is a bit transparent when in-game. Might have to do with sorting layers.